Move Settings from ubuntu_wizard to ubuntu_desktop_installer
