+++ packages/ubuntu_desktop_installer/lib/pages/try_or_install/try_or_install_page.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_html/flutter_html.dart';
 import 'package:provider/provider.dart';
-import 'package:ubuntu_wizard/settings.dart';
 import 'package:ubuntu_wizard/utils.dart';
 import 'package:ubuntu_wizard/widgets.dart';
 
 import '../../l10n.dart';
+import '../../settings.dart';
 import 'try_or_install_model.dart';
 
 export 'try_or_install_model.dart' show Option;
