+++ packages/ubuntu_desktop_installer/lib/settings.dart
