+++ packages/ubuntu_desktop_installer/pubspec.yaml
@@ -24,6 +24,7 @@ dependencies:
   form_field_validator: ^1.1.0
   glibc:
     path: ../glibc
+  gsettings: ^0.2.5
   intl: ^0.17.0
   nm: ^0.5.0
   provider: ^6.0.0
