+++ packages/ubuntu_desktop_installer/test/choose_your_look/choose_your_look_page_test.dart
@@ -4,7 +4,7 @@ import 'package:mockito/annotations.dart';
 import 'package:mockito/mockito.dart';
 import 'package:provider/provider.dart';
 import 'package:ubuntu_desktop_installer/pages/choose_your_look_page.dart';
-import 'package:ubuntu_wizard/settings.dart';
+import 'package:ubuntu_desktop_installer/settings.dart';
 import 'package:yaru_widgets/yaru_widgets.dart';
 
 import '../test_utils.dart';
