+++ packages/ubuntu_desktop_installer/test/choose_your_look/choose_your_look_page_test.mocks.dart
@@ -6,7 +6,7 @@ import 'dart:ui' as _i2;
 
 import 'package:flutter/material.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_wizard/settings.dart' as _i3;
+import 'package:ubuntu_desktop_installer/settings.dart' as _i3;
 
 // ignore_for_file: type=lint
 // ignore_for_file: avoid_redundant_argument_values
