+++ packages/ubuntu_desktop_installer/test/try_or_install/try_or_install_page_test.mocks.dart
@@ -7,7 +7,7 @@ import 'dart:ui' as _i2;
 
 import 'package:flutter/material.dart' as _i4;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_wizard/settings.dart' as _i3;
+import 'package:ubuntu_desktop_installer/settings.dart' as _i3;
 import 'package:ubuntu_wizard/src/utils/url_launcher.dart' as _i5;
 
 // ignore_for_file: type=lint
