+++ packages/ubuntu_wizard/pubspec.yaml
@@ -10,17 +10,14 @@ dependencies:
   args: ^2.2.0
   collection: ^1.15.0
   crypt: ^4.0.1
-  dbus: ^0.7.3
   diacritic: ^0.1.3
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
   form_field_validator: ^1.1.0
-  gsettings: ^0.2.5
   intl: ^0.17.0
   password_strength: ^0.2.0
-  provider: ^6.0.0
   safe_change_notifier: ^0.1.0
   subiquity_client:
     path: ../subiquity_client
